added one exception, but having issues with retrieving list of clothing categories
