+++ src/Shop/CouponException.java
@@ -1,0 +1,2 @@
+package Shop;public class CouponException {
+}
